Move RedisGraph to separate module
