+++ grapher/grapher/graph/__init__.py
@@ -1,0 +1,4 @@
+"""
+Graph storage abstraction layer
+"""
+from .redis import RedisGraph
